pyproject.toml
[project]
name = "pam-audio-extractor2"
version = "0.1.0"
description = "Passive acoustic monitoring audio analysis tool"
readme = "README.md"
requires-python = ">=3.13,<3.14" # use Python 3.13 because tensorflow v2.21.0 only has wheels for cp313
dependencies = [
    "nicegui>=3.8.0",
    "pywebview>=6.1.0",
    "platformdirs>=4.9.4",
    "psutil>=6.0.0",
    "guano>=1.0.16",
    "birdnet-analyzer>=2.4.0",
]

[project.scripts]
pam-analyzer = "pam_analyzer.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pam_analyzer"]  # required because of src layout

# Run ruff using uvx:
#  uvx ruff check src/pam_analyzer/
#  uvx ruff check src/pam_analyzer/ --fix
[tool.ruff]
line-length = 260
[tool.ruff.lint]
# Never enforce `E501` (line length violations).
ignore = ["E501"]
# Select specific error codes to enforce
select = [
    "F",    # Pyflakes (basic Python errors)
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "I",    # Import sorting
    "C4",   # Flake8 Comprehensions
    "B",    # Flake8 Bugbear (security and design checks)
    "UP",   # Pyupgrade (modernize Python code)
    "Q",    # flake8-quotes (enforce single quotes)
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"